back-end/requirements.txt
asgiref==3.2.7
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
coverage==5.1
Django==3.0.8
django-cors-headers==3.3.0
django-filter==2.2.0
djangorestframework==3.11.0
importlib-metadata==1.6.0
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
PyMySQL==0.10.0
pyparsing==2.4.7
pytest==5.4.2
pytest-cov==2.9.0
pytest-django==3.9.0
pytest-pythonpath==0.7.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
uWSGI==2.0.19.1
wcwidth==0.1.9
whitenoise==5.0.1
zipp==3.1.0
pyzbar==0.1.8
Pillow==7.2.0
